application/pyproject.toml
[project]
name = "fukushia-application"
version = "0.1.0"
description = "Memo & task management and assessment support app for social workers."
requires-python = ">=3.11"
dependencies = [
  "firebase-admin>=6.2.0",
  "google-generativeai>=0.3.0",
  "uvicorn[standard]>=0.35.0",
  "fastapi>=0.116.1",
  "langchain>=0.1.16",
  "langchain-google-genai>=0.1.3",
  "google-api-python-client>=2.86.0",
  "langchain-google-community>=0.0.4",
  "google-cloud-language>=2.17.2",
  "langgraph>=0.0.38",
  "google-genai>=1.32.0",
  "python-dotenv>=1.0.1",
  "pypdf2>=3.0.1",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
]

[build-system]
requires = ["hatchling>=1.12"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["app.py"]

[tool.ruff]
line-length = 120

[flake8]
max-line-length = 120

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]

[tool.hatch.scripts]
start = "python -m uvicorn application.main:app --reload --port 8000"
seed = "python application/scripts/seed.py"
